Force axp20x-rsb  module into initrd on ARM
On the aarch64 based Pine64 board it has been observed that the updated
dtb-allwinner package now requires the axp20x-regulator module in the
initrd for using the MMC driver, and its parent module axp20x-rsb does
not get loaded automatically even if added to the initrd.

https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/arch/arm64/boot/dts/allwinner/sun50i-a64-pine64.dts?id=3f241bfa60bdc9c4fde63fa6664a8ce00fd668c6

Without the module in initrd it will wait indefinitely for the root
partition on SD card on next boot.
Selecting an older kernel in GRUB will not help - only manually loading
an older dtb in U-Boot can help (if e.g. you had multiversion enabled).